1. I would like to clean the way futty handles session profiles (on disk/in registry).
Meaning of "Session from ..." ("storagetype" parameter) should not be
stored in the sessions file. Otherwise when you have sessions with
different storagetypes in them FuTTY behaivior becomes erratic and
difficult to explain.

This patch aims to simplify this a bit:

Since handling of putty.conf file is already there (based on patch from
Jakub Kotrla) I am adding "storagetype" parameter to putty.conf. Startup
behavior of all programs is modified to not depend implicitly of the
fact that original putty code only cares about registry, instead
explicit initialization order is used.

After the patch:

FuTTY will ALWAYS check for putty.conf (not only when "Sessions from
file" is selected). If putty.conf is not present - behave as before. If
it is present, but does not have "storagetype" present or "storagetype"
is set to "0" - behave as before. Otherwise select "Sessions from file"
behavior.

During run-time FuTTY will behave as before (but without strange changes
in the UI with radiobuttons dance) and it is now easy to predict where
stored sessions will be read from. Depending on the radiobutton selected
both storage destinations (file and registry) are available, however
destination is not saved - after FuTTY restart program behavior is again
defined by "storagetype" setting in putty.conf.

This allows "-load" switch in all programs to be dependent on the same
setting in putty.conf. In addition "-loadfile" will only look for stored
session in available files and new switch "-loadreg" will do the same
for registry. In cases when programs were looking for stored session in
both places (registy and then file) it will do the same, but order of
lookup will be dependent on "storagetype" setting. When "storagetype=1"
is specified in putty.conf session files will be searched before
registry.

It should be relativly easy now to actually store "storagetype" in
putty.conf based on radiobutton selection, however I think this is not
really necessary.

I fixed some (not all) border conditions in putty.conf parsing routine,
so it would not segfault when last character in the file is not "\n".

2. Fixed wincons.c to properly handle situation when plink is being used
by another program (for example in cases with subwersion "ssh+svn"
protocol) - use proper console device rather than StdHandle (from Yahor
Shybeka).